Add showMatrix
